Remove location methods for ZNS and ENS
