+++ src/Ens.ts
@@ -10,7 +10,6 @@ import {ResolutionError, ResolutionErrorCode} from './errors/resolutionError';
 import EthereumContract from './contracts/EthereumContract';
 import EnsNetworkMap from 'ethereum-ens-network-map';
 import {
-  BlockchainType,
   EnsSource,
   Locations,
   NamingServiceName,
@@ -29,7 +28,6 @@ import ConfigurationError, {
 } from './errors/configurationError';
 import {EthereumNetworks} from './utils';
 import {requireOrFail} from './utils/requireOrFail';
-import {GetLocations} from './utils/LocationUtils';
 
 /**
  * @internal
@@ -46,7 +44,6 @@ export default class Ens extends NamingService {
   readonly url: string | undefined;
   readonly provider: Provider;
   readonly readerContract: EthereumContract;
-  readonly blockchain: BlockchainType = BlockchainType.ETH;
 
   constructor(
     source: EnsSource = {
@@ -231,7 +228,10 @@ export default class Ens extends NamingService {
   }
 
   locations(domains: string[]): Promise<Locations> {
-    return GetLocations(domains, this);
+    throw new ResolutionError(ResolutionErrorCode.UnsupportedMethod, {
+      method: NamingServiceName.ENS,
+      methodName: 'locations',
+    });
   }
 
   /**
